Initial split of files
