+++ config/default.json
@@ -1,0 +1,9 @@
+{
+  "debug": true,
+  "from": "D:/Unsorted Torrents/2017-torrents/",
+  "temp": "D:/Sorted Torrents/temp",
+  "to": "D:/Sorted Torrents/TV/",
+  "videoExtensions": ["mkv", "avi", "mp4"],
+  "ignorePatterns": ["RARBG.mp4", "*sample*", "*incomplete*"],
+  "filebotCommand": "filebot -rename \"$PATTERN$\" --db TheTVDB -non-strict"
+}
