+++ package.json
@@ -9,6 +9,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "config": "^2.0.1",
     "recursive-readdir": "^2.2.1"
   },
   "devDependencies": {
